KUBIG CONFERENCE/이 민원은 접수되었습니다/requirements.txt
langchain==0.3.2
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.8
langchain-upstage==0.3.0
langgraph==0.2.34
python-dotenv==1.0.1
ipython==8.18.1
gradio==4.44.1
gdown==5.2.0
langchain-text-splitters==0.3.0
pandas==2.2.2
selenium==4.25.0
webdriver-manager==4.0.2
requests==2.32.3
chromedriver-autoinstaller==0.6.4
requests==2.32.3